requirements.txt
opencv-python>=4.9.0
numpy>=1.26
Pillow>=10.0
svgwrite>=1.4.3
svgpathtools>=1.6.1
scikit-image>=0.22
PyYAML>=6.0
cairosvg>=2.7.1
tkinterdnd2>=0.3.0
